refactor(ledger): collocate the portfolio route registrars :hammer:
Both portfolio registrars move out of portfolio_handler.go into
portfolio_routes.go, leaving the handler file with the handler struct, the
request/response envelopes and the cores.

No route, authz tuple, operation ID or path changed.
